refactor(otel): centralize provider-mode selection in config via ProviderSource
Move the ProviderSource enum into the config module and add a single
resolveProviderSource(config) resolver that owns the precedence chain
(explicit tracerProvider > useDefaultTracerProvider > auto-OTLP).

- createTracerProvider now resolves the source once and switches on it,
  instead of re-deriving the tiers from raw config booleans inline.
- ExecutionOtelPlugin stores providerSource and gates the child
  Invocation span on 'source !== Global' (was '!useDefaultTracerProvider').
- InvocationOtelPlugin drops its write-only useDefaultTracerProvider field.
- provider.ts re-exports ProviderSource so existing import sites are stable.

No behavior change. 194/194 otel tests pass; tsc --noEmit + build clean.
